requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
google-genai>=0.8.0
google-api-python-client>=2.150.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
duckduckgo-search>=6.2.0
sse-starlette>=2.1.0
youtube-transcript-api>=0.6.0
